Arithmetics are printed. It also includes negative Ints and Floats support. Negative constants are also added to the Function Table

+++ Scanner_Parser/Parser_Structures/SemanticCube.py
@@ -28,7 +28,7 @@ class SemanticCube:
             ("Int", "+", "Int") : "Int",
             ("Int", "-", "Int"): "Int",
             ("Int", "*", "Int"): "Int",
-            ("Int", "/", "Int"): "Float",
+            ("Int", "/", "Int"): "Int",
             ("Int", "%", "Int"): "Int",
 
             # Int vs Float
